Extract streaming_reply.rs from turn.rs (Phase 5, #685)
Issue #685 (parent #680, Phase 5) — single-PR migration since the
responsibility is small (single struct + 2 helpers + 2 free fns).
Create the new sibling module `src/agent/loop_run/streaming_reply.rs`
and move the streaming-reply render state plus the chunk-handling
flow out of `turn.rs`.

Migrated to `streaming_reply.rs` (`pub(super)`):

* `struct StreamingReplyRenderState { first_chunk, renderer }`
* `impl StreamingReplyRenderState { new() }` (markdown disabled →
  `renderer = None`; otherwise a fresh `MarkdownRenderer` per stream)
* `streaming_reply_needs_prefix(first_chunk, stream_output) -> bool`
* `streaming_reply_needs_trailing_newline(first_chunk, stream_output)
    -> bool`
* `handle_streaming_assistant_chunk(render_state, chunk, stream_output,
    stop_signal, interrupt_flag) -> Result<(), String>`
  (formerly `impl Agent::handle_streaming_assistant_chunk`, an
  associated fn that does not access `self`; converted to free fn)
* `finish_streaming_assistant_reply(render_state, stream_output)`
  (same pattern)

`turn.rs` adjustments:

* Removed the struct + impl + 2 helper fns + 2 associated fns
  (originally at lines 2273-2303 and 5534-5582).
* `stream_assistant_reply` (the only caller in `impl Agent`) now
  calls `super::streaming_reply::StreamingReplyRenderState::new()`
  and `super::streaming_reply::handle_streaming_assistant_chunk(...)`
  / `finish_streaming_assistant_reply(...)` directly.
* Test mod call sites (`super::streaming_reply_needs_{prefix,trailing_newline}`)
  re-routed to `super::super::streaming_reply::*`.

`loop_run.rs` adjustments:

* Registered `mod streaming_reply;` with a Phase-5 / Issue-685 doc
  comment.

`CLAUDE.md` File Map updated:

* Refreshed the `reminder_pipeline.rs` entry to reflect the
  consolidated state after Phase 4 close (batches 1-3 complete, no
  veneer shims remaining on `impl Agent`).
* Added the new `streaming_reply.rs` entry.

LOC delta:

* `turn.rs`: 29,995 → 29,891 (-104 LOC)
* `streaming_reply.rs`: 0 → 104 (new)
* Cumulative `turn.rs`: 39,489 → 29,891 (-9,598 LOC, -24.3%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
