+++ CLAUDE.md
@@ -15,7 +15,8 @@ Rust 版 local-first coding agent (Ollama 専用、`workspace/v0.1.0` ベース)
 - `src/agent/loop_run/actor_loop_flow.rs` — Issue #681 (parent #680, Phase 1): actor loop 制御フローのデータ型を `turn.rs` から抽出。`PostReplyRecoveryArgs/Outcome` + 27 個の `ActorLoop*Args` / `ActorLoop*Outcome` struct/enum (pre-reply / tool-preparation / no-tool / task-contract-reply / completion / post-tool / missing-repo-change / empty-reply / prose-only / plan-tool-followup) + `TaskContractVerifierFlowOutcome` + 2 つの Outcome constructor (`repair_job_done_outcome` / `missing_repo_edits_finalize_outcome`) を保持。`pub(super)` 限定、facade 再 export なし (DR3-001)、`turn.rs` が唯一の in-crate consumer。`impl Agent` method の free-function 化は後続コミットで漸進実施 (既存 "Split X helpers" 粒度)。
 - `src/agent/loop_run/verifier_orchestration.rs` — Issue #682 (parent #680, Phase 2): verifier orchestration の 7 データ型を `turn.rs` から抽出。`JobInstallOutcome` (PR-001 SSOT) / `VerifierDiagnosticPassOutcome` (Accepted / RetryPending / Unavailable / Skipped) / `PreparedVerifierDiagnosticPass` / `PreparedVerifierRepairPass` / `VerifierRepairAttemptProgress` (Return / Continue / Break trichotomy) / `StructuredTaskContractVerifierRun` / `TaskContractVerifierFlowArgs<'a, 'b>` を保持。`pub(super)` 限定、facade 再 export なし (DR3-001)。`turn.rs` 内 dispatch method (`run_verifier_diagnostic_pass` / `verifier_repair_pass` / `drive_task_contract_verifier` / `record_controller_verifier_repair_invalid` / `emit_safe_stop_report_for_repair_exhausted`) は後続 PR で漸進的に migrate (Phase 1 / actor_loop_flow パターン踏襲)。`actor_loop_flow.rs` は `super::verifier_orchestration::TaskContractVerifierFlowArgs` 経由で参照。
 - `src/agent/loop_run/scaffold_pipeline.rs` — Issue #683 (parent #680, Phase 3): scaffold / deterministic-scaffold pipeline データ型 + telemetry event 名定数 + 全 dispatch method を `turn.rs` から抽出 (batches 1-17 完了)。`ScaffoldFramework` (Next / React / Nuxt) + `impl { label, scaffold_hint }` / `PlanExplorationKey` (stage / tool_name / normalized_args triple、`actor_loop_flow` の repeated-exploration detector 用) / `ScaffoldFallbackResult` (NotApplicable / Applied / Failed / Skipped) / `DeterministicScaffoldSpec` (label / event / scaffold_kind / files) / `ScaffoldDiffStatus` (NotScaffold / Changed / UnchangedOrMissing) を保持。`EVENT_DETERMINISTIC_FASTAPI_SCAFFOLD` / `_PYTHON_CLI` / `_FORMAT_ERROR_SMALL_EDIT` / `_PYTHON_TEST_FALLBACK` / `CREATE_NEXT_APP_PACKAGE_VERSION` を `pub(super) const` で SSOT 化。全 scaffold dispatch fn (`maybe_apply_deterministic_nextjs_scaffold` / `maybe_materialize_{mode,task_contract,framework_game,python_test}_fallback` / `maybe_apply_deterministic_{quality,polish}_fallback` / `_after_timeout` / `maybe_apply_local_llm_small_edit_fallback` / `materialize_deterministic_fallback_plan` / `maybe_apply_deterministic_edit_after_format_error` / `maybe_fallback_plan_model_after_timeout` / `maybe_apply_requested_port_script` / `maybe_apply_deterministic_framework_support_files` / `empty_workspace_scaffold_policy_error` / `post_scaffold_{edit,continuation}_recovery_{message,target}` / `repo_edit_has_post_scaffold_delta` / `scaffold_candidate_for_missing_role` / `active_task_{requires_nextjs_scaffold,requested_scaffold_framework}` / `python_csv_names_from_request_and_anvil` / `mode_deterministic_scaffold_spec` / `write_deterministic_scaffold_files` / `finalize_deterministic_scaffold_materialization` / `record_scaffold_artifact_snapshot` / `deterministic_nextjs_scaffold_skip_reason` ほか) を `pub(super) fn name(agent: &/&mut Agent, ...)` 形式で保持。caller は `turn.rs` / `actor_loop_flow.rs` / `verifier_orchestration.rs` / 本 module 内の sibling 呼び出し (`super::scaffold_pipeline::*`)。`pub(super)` 限定、facade 再 export なし (DR3-001)。`impl Agent` には scaffold veneer shim を残さない。
-- `src/agent/loop_run/reminder_pipeline.rs` — Issue #684 (parent #680, Phase 4): reminder dispatch orchestration を `turn.rs` から抽出。`ReminderCallContext` (session_id / model / kind / frame / mode_label / active_precautions_summary / touched_files / user_task / workspace_root / active_precautions_at_call_time / anvil_score + `_from_current_turn` flag) struct + `impl { inputs() -> ReminderInputs<'_>, log_outcome(turn_index, &outcome, include_inputs) }` を保持。`reminder.rs` sibling の public SSOT (`ReminderInputs` / `ReminderOutcome` / `build_log_payload`) は変更しない (orchestration が caller を assemble する経路だけを切り出す)。`pub(super)` 限定、facade 再 export なし (DR3-001)。dispatch method (`maybe_invoke_reminder` / `prepare_reminder_context` / `try_clone_reminder_client` / `run_reminder_sidecar` ほか) は `turn.rs` に残し、後続 PR で漸進的に migrate (Phase 3 / scaffold_pipeline パターン踏襲)。
+- `src/agent/loop_run/reminder_pipeline.rs` — Issue #684 (parent #680, Phase 4): reminder dispatch orchestration を `turn.rs` から完全抽出 (batches 1-3 完了)。`ReminderCallContext` struct + `impl { inputs() -> ReminderInputs<'_>, log_outcome(turn_index, &outcome, include_inputs) }` + 全 dispatch fn (`maybe_invoke_reminder` / `reminder_feedback_kind` / `reminder_gate` / `prepare_reminder_context` / `reminder_mode_label` / `active_precautions_at_call_time` / `try_clone_reminder_client` / `run_reminder_sidecar`) を `pub(super) fn name(agent: &/&mut Agent, ...)` 形式で保持。caller は `turn.rs` (内部メソッド経路なし) / `actor_loop_flow.rs` (`super::reminder_pipeline::maybe_invoke_reminder(agent, ...)` 直呼出)。`reminder.rs` sibling の public SSOT (`ReminderInputs` / `ReminderOutcome` / `build_log_payload`) は変更しない。`pub(super)` 限定、facade 再 export なし (DR3-001)。`impl Agent` には reminder veneer shim を残さない。
+- `src/agent/loop_run/streaming_reply.rs` — Issue #685 (parent #680, Phase 5): streaming reply render state + chunk-handling flow を `turn.rs` から抽出。`StreamingReplyRenderState { first_chunk, renderer: Option<MarkdownRenderer> }` struct + `impl { new() }` (markdown disabled 時 `renderer = None` / TTY color + utf-8 capability 検査) + `streaming_reply_needs_prefix` / `streaming_reply_needs_trailing_newline` 純関数 predicate + `handle_streaming_assistant_chunk(render_state, chunk, stream_output, stop_signal, interrupt_flag)` / `finish_streaming_assistant_reply(render_state, stream_output)` (旧 `impl Agent` 関連 fn を free fn 化、`&Agent` 不要) を保持。`turn.rs::stream_assistant_reply` が `super::streaming_reply::*` 直接呼出で消費。`pub(super)` 限定、facade 再 export なし (DR3-001) — `turn.rs` が唯一の in-crate consumer。
 - `src/agent/loop_run/project_verifier.rs` — `ProjectVerifier` capability (verifier_repair cheap check SSOT, Issue #639). private mod, not re-exported (DR3-001) — `turn.rs` is the only in-crate consumer.
 - `src/agent/loop_run/semantic_failure.rs` — Issue #647 Phase A.1: `SemanticFailureReport` schema + `build_failure_cluster_from_observation` + `cluster_key` (deterministic, sanitize-then-shape-then-hash). private mod, not re-exported (DR3-001).
 - `src/agent/loop_run/spec_authority.rs` — Issue #647 Phase A.2: `SpecAuthority` enum (5 variant) + `select_authority` + `ArtifactConsensus` tie-break + `detect_test_weakening` / `detect_impl_weakening` (9 pure-fn detectors). private mod, not re-exported (DR3-001).
